requirements.txt
tencentcloud-sdk-python
qiniu
cryptography
requests